requirements.txt
asgiref
certifi
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
Markdown
MouseInfo
mysql-connector-python
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
typing_extensions
tzdata
virtualenv
